refactor: remove re-exports of instrumentation imports
These aren't particularly necessary, are currently causing issues with
the click CLI autocomplete times due to the modules being loaded each
time, and will potentially long-term cause circular import issues. Let's
just use the full paths for imports instead here.
