+++ src/dvsim/instrumentation/report/__init__.py
@@ -3,17 +3,3 @@
 # SPDX-License-Identifier: Apache-2.0
 
 """DVSim Scheduler Instrumentation report."""
-
-from dvsim.instrumentation.report.base import (
-    InstrumentationVisualizer,
-    RenderProfile,
-    render_html_report,
-)
-from dvsim.instrumentation.report.registry import ReportVisualizationRegistry
-
-__all__ = (
-    "InstrumentationVisualizer",
-    "RenderProfile",
-    "ReportVisualizationRegistry",
-    "render_html_report",
-)
