+++ src/dvsim/instrumentation/report/usage.py
@@ -10,8 +10,11 @@ from collections.abc import Callable
 import plotly.graph_objects as go
 from plotly.graph_objs import Figure
 
-from dvsim.instrumentation import InstrumentationResults
-from dvsim.instrumentation.records import ConcreteJobTimingMetrics, JobInstrumentationResults
+from dvsim.instrumentation.records import (
+    ConcreteJobTimingMetrics,
+    InstrumentationResults,
+    JobInstrumentationResults,
+)
 from dvsim.instrumentation.report.base import (
     DEFAULT_VISUALIZATION_HEIGHT_PX,
     PLOTLY_TIMING_AXIS_CONFIG,
